Fix URL unwrap fatal error in xcode preview
Updated FileManager extension to use init(fileURLWithPath:)
URL initializer when constructing file path in switch statement
for 'file' and 'directory' types.

Xcode preview was crashing due to the URL force unwrap encountering
a nil value. This happens when the path contains a space and hence
URL(string:) returns a nil value.
The root cause; "Simulator Devices" folder located in:
/Users/<username>/Library/Developer/Xcode/UserData/Previews

On my device i actuall have a directory with a space and a directory
with a URL encoded space;
./Simulator Devices
./Simulator%20Devices

Making the above change prevents the fatal error and still seems to use
the correct directory (Simulator Devices).
I imagine this is really an apple/xcode issue and there should only
be one of those directories present.

Tested on Simulator and physical device; working as expected.